XHTML compatible Entity for ×
This changes &times; to &#215; (the "close growl" symbol in this case).

The entity &times; is not a valid entity for an XHTML document – and opening a growl would produce an error while trying to attach the growl to the document. The entity &#215; is the numerical representation of &times;, and can be used in HTML and XHTML documents. This makes bootstrap-growl wider usable.